update author services responses

+++ src/main/java/com/btv/app/features/author/services/AuthorController.java
@@ -1,13 +1,17 @@
 package com.btv.app.features.author.services;
 
 import com.btv.app.features.author.model.Author;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
-@RequestMapping("api/author")
+@RequestMapping("api/admin")
 public class AuthorController {
     private final AuthorService authorService;
 
@@ -16,54 +20,62 @@ public class AuthorController {
     }
 
     @GetMapping("/getAllAuthors")
-    public List<Author> getAllAuthors(){
+    public ResponseEntity<List<Author>>getAllAuthors(){
         try {
-            return authorService.getAllAuthors();
+            List<Author> res =  authorService.getAllAuthors();
+            return ResponseEntity.status(200).body(res);
         } catch (Exception e) {
-            System.out.println(e);
-            return null;
+            return ResponseEntity.status(500).build();
         }
     }
 
     @GetMapping("/getAuthorByID/{id}")
-    public Author getAuthorByID(@PathVariable("id") Long id){
+    public ResponseEntity<Author> getAuthorByID(@PathVariable("id") Long id){
         try{
-            return authorService.getAuthorByID(id);
+            Author res = authorService.getAuthorByID(id);
+            return ResponseEntity.status(200).body(res);
         } catch (Exception e) {
-            System.out.println(e);
-            return null;
+            return ResponseEntity.status(500).build();
         }
     }
 
     @PostMapping("/addAuthor")
-    public Author addAuthor(@ModelAttribute Author author) {
+    public ResponseEntity<Author> addAuthor(@ModelAttribute Author author) {
         try{
-            Author tmp = authorService.addAuthor(author);
-            return new Author(tmp);
+            Author res = authorService.addAuthor(author);
+            return ResponseEntity.status(200).body(res);
         } catch (Exception e) {
-            System.out.println(e);
-            return null;
+            return ResponseEntity.status(500).build();
         }
     }
 
+
     @PutMapping("/modifyAuthor/{id}")
-    public Author modifyAuthor(@PathVariable("id") Long id, @ModelAttribute Author author){
+    public ResponseEntity<Author> modifyAuthor(@PathVariable("id") Long id, @ModelAttribute Author author){
         try{
             Author curAuthor = authorService.getAuthorByID(id);
+            if(curAuthor == null){
+                return ResponseEntity.status(404).build();
+            }
             Author tmp = authorService.modifyAuthor(curAuthor, author);
-            return new Author(tmp);
+            Author res = new Author(tmp);
+            return ResponseEntity.status(200).body(res);
         } catch (Exception e) {
-            System.out.println(e);
-            return null;
+            return ResponseEntity.status(500).build();
         }
     }
 
     @DeleteMapping("/deleteAuthor/{id}")
-    public void deleteAuthor(@PathVariable("id") Long id){
+    public ResponseEntity<Author> deleteAuthor(@PathVariable("id") Long id){
         try{
+            Author curAuthor = authorService.getAuthorByID(id);
+            if(curAuthor == null){
+                return ResponseEntity.status(404).build();
+            }
             authorService.deleteAuthor(id);
+            return ResponseEntity.status(200).body(curAuthor);
         } catch (Exception e) {
-            System.out.println(e);
+            return ResponseEntity.status(500).build();
         }
     }
 }
